test: 서버사이드 렌더링 시 테스트 페이지 생성
